refactor: 폴더 구조 변경, styled components 파일 분리
