+++ components/community/Board/ControlBar/SearchInputBar/index.tsx
@@ -1,0 +1,7 @@
+import * as S from "./style";
+
+const BoardSearchInput = () => {
+  return <S.InputBar />;
+};
+
+export default BoardSearchInput;
