+++ components/community/Board/ControlBar/style.ts
@@ -1,0 +1,30 @@
+import { styled } from "styled-components";
+
+export const ControlBarBox = styled.div`
+  width: 100%;
+
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 30px;
+`;
+
+export const SearchBar = styled.div`
+  position: relative;
+
+  display: flex;
+  align-items: center;
+`;
+
+export const ControlMenuContainer = styled.div`
+  position: absolute;
+  left: 4px;
+`;
+
+export const SearchIconContainer = styled.button`
+  position: absolute;
+  right: 16px;
+`;
+
+export const WriteButtonContainer = styled.button`
+  padding-left: 130px;
+`;
