+++ components/community/Board/New/Editor/ContentInput/index.tsx
@@ -1,0 +1,23 @@
+import { SetStateAction, useState } from "react";
+import { useRecoilState, useSetRecoilState } from "recoil";
+import { postContentState } from "recoil/states";
+
+import * as S from "./style";
+
+const EditorContentInput = () => {
+  const [content, setContent] = useRecoilState(postContentState);
+
+  return (
+    <S.Box>
+      <S.Text>내용</S.Text>
+      <S.Input
+        value={content}
+        onChange={(e: { target: { value: SetStateAction<string> } }) =>
+          setContent(e.target.value)
+        }
+      ></S.Input>
+    </S.Box>
+  );
+};
+
+export default EditorContentInput;
