+++ components/community/Board/New/Editor/ContentInput/style.ts
@@ -1,0 +1,28 @@
+import { styled } from "styled-components";
+
+export const Box = styled.div``;
+
+export const Text = styled.span`
+  padding-left: 12px;
+  background: #f7f7f9;
+  color: #000000;
+
+  font-weight: 600;
+  font-size: 18px;
+  line-height: 21px;
+
+  display: flex;
+  align-items: center;
+
+  height: 45px;
+`;
+
+export const Input = styled.textarea`
+  resize: none;
+
+  width: 1075px;
+  height: 477px;
+  border-radius: 0px 0px 30px 30px;
+  font-size: 18px;
+  padding: 12px;
+`;
