+++ components/community/Board/New/Editor/TitleInput/style.ts
@@ -1,0 +1,34 @@
+import styled from "styled-components";
+
+export const Box = styled.div`
+  display: flex;
+  align-items: center;
+
+  padding-bottom: 45px;
+
+  gap: 25px;
+`;
+
+export const Text = styled.h1`
+  font-weight: 600;
+  font-size: 18px;
+  line-height: 21px;
+  display: flex;
+  align-items: center;
+  text-align: center;
+
+  color: #000000;
+`;
+
+export const Input = styled.input`
+  box-sizing: border-box;
+  width: 984px;
+  height: 66px;
+  font-size: 18px;
+
+  padding-left: 16px;
+
+  background: #ffffff;
+  border: 1px solid #f7f7f9;
+  border-radius: 30px;
+`;
