+++ components/community/Board/New/Editor/index.tsx
@@ -1,0 +1,15 @@
+import TitleInput from "./TitleInput";
+import ContentInput from "./ContentInput";
+
+import * as S from "./style";
+
+const Editor = () => {
+  return (
+    <S.Box>
+      <TitleInput />
+      <ContentInput />
+    </S.Box>
+  );
+};
+
+export default Editor;
