+++ components/community/Board/PostListBox/style.ts
@@ -1,0 +1,7 @@
+import { styled } from "styled-components";
+
+export const Box = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 45px;
+`;
