+++ components/community/Board/style.ts
@@ -1,0 +1,10 @@
+import { styled } from "styled-components";
+
+export const Box = styled.div`
+  width: 100%;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
