--- components/community/BulletinBoard/index.tsx
@@ -1,8 +1,0 @@
-/** community - 게시판 */
-const BulletinBoard = () => {
-  return (
-    <p>게시판</p>
-  );
-};
-
-export default BulletinBoard;
