+++ pages/community/board/new/index.tsx
@@ -1,0 +1,7 @@
+import New from "components/community/Board/New";
+
+const Index = () => {
+  return <New />;
+};
+
+export default Index;
