+++ pages/community/friends/index.tsx
@@ -1,10 +1,8 @@
-import Friends from 'components/community/Friends';
+import Friends from "components/community/Friends";
 
 /** community - 친구 페이지 */
 const Index = () => {
-  return (
-    <Friends />
-  );
+  return <Friends />;
 };
 
 export default Index;
