--- pages/community/posts/index.tsx
@@ -1,9 +1,0 @@
-import BulletinBoard from 'components/community/BulletinBoard';
-
-const CommunityPostsPage = () => {
-  return (
-    <BulletinBoard />
-  );
-};
-
-export default CommunityPostsPage;
